requirements/req.txt
django-taggit
django-registration


docutils
pygments
